+Changes: modified train.py. Still needs to fix the issue of overfitting to center

+++ NAS/dataset.py
@@ -15,8 +15,8 @@ class ImageDataset(Dataset):
             all_images.append(torch.from_numpy(images))  # (N, 1, H, W)
             all_turns.append(torch.from_numpy(turns))    # (N,)
 
-        self.images = torch.cat(all_images, dim=0).to(torch.uint8)  # shape: (Total, 1, H, W)
-        self.turns = torch.cat(all_turns, dim=0).to(torch.int8)     # shape: (Total,)
+        self.images = torch.cat(all_images, dim=0).to(torch.float32)  # shape: (Total, 1, H, W)
+        self.turns = torch.cat(all_turns, dim=0).to(torch.float32)     # shape: (Total,)
 
     def __len__(self):
         return len(self.turns)
